Creado fichero para funciones relacionadas con enviar mensajes a los jugadores
